lib/py: Automatically decode immutable struct fields as immutables.
Currently, Thrift requires exhaustive immutable annotations for all
container fields of an immutable struct to work correctly:

    struct ImmutableParent {
      1: list<string> (python.immutable="") str_list,
      2: set<list<string> (python.immutable="")> (python.immutable="") set_of_lists,
      3: map<string,list<string> (python.immutable="")> (python.immutable="")map_of_lists,
    } (
      python.immutable = "",
    )

This is bad for several reasons:

    1. Bad ergonomics, as definitions quickly get polluted by the
       annotations, and it is easy to miss.

    2. Danger of missed annotation causing the decoder to fail
       with TypeError('unhashable type').

    3. Lack of proper support for nested annotations in third-party
       compilers like Scrooge.

This diff changes the handling of immutable structs in the Python
decoder to automatically assume that all container fields of
immutable structs are also immutable.  This allows to side-step all
problems mentioned above.

This commit makes no attempt to validate nesting of mutable structs
within immutable structs, as that is a separate concern.